+++ ssqueezepy/_ssq_stft.py
@@ -1,15 +1,14 @@
 # -*- coding: utf-8 -*-
 import numpy as np
-from ._stft import stft, phase_stft, get_window, _check_NOLA
+from ._stft import stft, get_window, _check_NOLA
 from ._ssq_cwt import _invert_components, _process_component_inversion_args
-from .utils import WARN
-from .ssqueezing import ssqueeze
-
-pi = np.pi
+from .utils import WARN, EPS, pi
+from .ssqueezing import ssqueeze, _check_squeezing_arg
 
 
-def ssq_stft(x, window=None, n_fft=None, win_len=None, hop_len=1,
-             fs=1., padtype='reflect', modulated=True, squeezing='sum'):
+def ssq_stft(x, window=None, n_fft=None, win_len=None, hop_len=1, fs=1.,
+             modulated=True, ssq_freqs=None, padtype='reflect', squeezing='sum',
+             mapkind='maximal', gamma=None):
     """Synchrosqueezed Short-Time Fourier Transform.
     Implements the algorithm described in Sec. III of [1].
 
@@ -20,16 +19,26 @@ def ssq_stft(x, window=None, n_fft=None, win_len=None, hop_len=1,
         window, n_fft, win_len, hop_len, fs, padtype, modulated
             See `help(stft)`.
 
-        squeezing: str
-            Synchrosqueezing scheme to use. See `help(ssqueeze)`.
+        ssq_freqs, squeezing, mapkind
+            See `help(ssqueezing.ssqueeze)`.
+
+        gamma: float / None
+            STFT phase threshold. Sets `w=inf` for small values of `Sx` where
+            phase computation is unstable and inaccurate (like in DFT):
+                w[abs(Sx) < beta] = inf
+            This is used to zero `Sx` where `w=0` in computing `Tx` to ignore
+            contributions from points with indeterminate phase.
+            Default = sqrt(machine epsilon) = np.sqrt(np.finfo(np.float64).eps)
 
     # Returns:
         Tx: np.ndarray
-            Synchrosqueezed STFT of `x`, of same shape as `Sx` (see `stft`).
+            Synchrosqueezed STFT of `x`, of same shape as `Sx`.
         ssq_freqs: np.ndarray
-            Frequencies associated with rows of `Tx` and `Sx`.
+            Frequencies associated with rows of `Tx`.
         Sx: np.ndarray
             STFT of `x`. See `help(stft)`.
+        Sfs: np.ndarray
+            Frequencies associated with rows of `Sx` (by default == `ssq_freqs`).
         dSx: np.ndarray
             Time-derivative of STFT of `x`. See `help(stft)`.
         w: np.ndarray
@@ -41,13 +50,14 @@ def ssq_stft(x, window=None, n_fft=None, win_len=None, hop_len=1,
         https://arxiv.org/abs/1006.2533
     """
     n_fft = n_fft or len(x)
+    _check_squeezing_arg(squeezing)
 
     Sx, dSx = stft(x, window, n_fft=n_fft, win_len=win_len,
                    hop_len=hop_len, fs=fs, padtype=padtype,
                    modulated=modulated, derivative=True)
 
     Sfs = np.linspace(0, .5, Sx.shape[0]) * fs
-    w = phase_stft(Sx, dSx, Sfs)
+    w = phase_stft(Sx, dSx, Sfs, gamma)
 
     Tx, ssq_freqs = ssqueeze(Sx, w, transform='stft', squeezing=squeezing,
                              ssq_freqs=Sfs)
@@ -56,7 +66,7 @@ def ssq_stft(x, window=None, n_fft=None, win_len=None, hop_len=1,
 
 
 def issq_stft(Tx, window=None, cc=None, cw=None, n_fft=None, win_len=None,
-              hop_len=1, N=None, modulated=True):
+              hop_len=1, modulated=True):
     """Inverse synchrosqueezed STFT.
 
     # Arguments:
@@ -111,3 +121,52 @@ def issq_stft(Tx, window=None, cc=None, cw=None, n_fft=None, win_len=None,
 
     x *= (2 / window[len(window)//2])
     return x
+
+
+def phase_stft(Sx, dSx, Sfs, gamma=None):
+    """Phase transform of STFT:
+        w[u, k] = Im( k - d/dt(Sx[u, k]) / Sx[u, k] / (j*2pi) )
+
+    # Arguments:
+        Sx: np.ndarray
+            STFT of `x`, where `x` is 1D.
+
+        dSx: np.ndarray
+            Time-derivative of STFT of `x`
+
+        Sfs: np.ndarray
+            Associated physical frequencies, according to `dt` used in `stft`.
+            Spans 0 to fs/2, linearly.
+
+        gamma: float / None
+            STFT phase threshold. Sets `w=inf` for small values of `Sx` where
+            phase computation is unstable and inaccurate (like in DFT):
+                w[abs(Sx) < beta] = inf
+            This is used to zero `Wx` where `w=0` in computing `Tx` to ignore
+            contributions from points with indeterminate phase.
+            Default = sqrt(machine epsilon) = np.sqrt(np.finfo(np.float64).eps)
+
+    # Returns:
+        w: np.ndarray
+            Phase transform for each element of `Sx`. w.shape == Sx.shape.
+
+    # References:
+        1. Synchrosqueezing-based Recovery of Instantaneous Frequency from
+        Nonuniform Samples. G. Thakur and H.-T. Wu.
+        https://arxiv.org/abs/1006.2533
+
+        2. The Synchrosqueezing algorithm for time-varying spectral analysis:
+        robustness properties and new paleoclimate applications.
+        G. Thakur, E. Brevdo, N.-S. Fučkar, and H.-T. Wu.
+        https://arxiv.org/abs/1105.0010
+    """
+    w = Sfs.reshape(-1, 1) - np.imag(dSx / Sx) / (2*pi)
+
+    # treat negative phases as positive; these are in small minority, and
+    # slightly aid invertibility (as less of `Wx` is zeroed in ssqueezing)
+    w = np.abs(w)
+
+    gamma = gamma or np.sqrt(EPS)
+    # threshold out small points
+    w[np.abs(Sx) < gamma] = np.inf
+    return w
