+++ ssqueezepy/ssqueezing.py
@@ -3,14 +3,13 @@ import numpy as np
 from types import FunctionType
 from .algos import find_closest, indexed_sum, replace_at_inf
 from .utils import process_scales, _infer_scaletype, _process_fs_and_t
-from .utils import NOTE, EPS, pi
+from .utils import NOTE, pi
 from .wavelets import center_frequency
 
 
 def ssqueeze(Wx, w, ssq_freqs=None, scales=None, fs=None, t=None, transform='cwt',
              squeezing='sum', mapkind='maximal', wavelet=None):
-    """Synchrosqueezes the CWT or STFT of `x`. Used internally by `synsq_cwt`
-    and `synsq_stft_fwd`.
+    """Synchrosqueezes the CWT or STFT of `x`.
 
     # Arguments:
         Wx or Sx: np.ndarray
@@ -61,11 +60,11 @@ def ssqueeze(Wx, w, ssq_freqs=None, scales=None, fs=None, t=None, transform='cwt
             must match `Wx.shape`.
 
         mapkind: str['maximal', 'peak', 'energy']
-            See help(ssq_cwt).
+            See `help(ssq_cwt)`.
 
         wavelet: wavelets.Wavelet
             Only used if mapkind != 'maximal' to compute center frequencies.
-            See help(cwt).
+            See `help(cwt)`.
 
     # Returns:
         Tx: np.ndarray [nf x n]
@@ -158,14 +157,7 @@ def ssqueeze(Wx, w, ssq_freqs=None, scales=None, fs=None, t=None, transform='cwt
         if transform not in ('cwt', 'stft'):
             raise ValueError("`transform` must be one of: cwt, stft "
                              "(got %s)" % squeezing)
-
-        supported = ('sum', 'lebesgue', 'abs')
-        if not isinstance(squeezing, (str, FunctionType)):
-            raise TypeError("`squeezing` must be string or function "
-                            "(got %s)" % type(squeezing))
-        elif isinstance(squeezing, str) and squeezing not in supported:
-            raise ValueError("`squeezing` must be one of: sum, lebesgue, abs "
-                              "(got %s)" % ', '.join(supported))
+        _check_squeezing_arg(squeezing)
 
         if scales is None and transform == 'cwt':
             raise ValueError("`scales` can't be None if `transform == 'cwt'`")
@@ -209,153 +201,11 @@ def ssqueeze(Wx, w, ssq_freqs=None, scales=None, fs=None, t=None, transform='cwt
     return Tx, ssq_freqs
 
 
-def phase_cwt(Wx, dWx, difftype='direct', gamma=None):
-    """Calculate the phase transform at each (scale, time) pair:
-          w[a, b] = Im((1/2pi) * d/db (Wx[a,b]) / Wx[a,b])
-    See above Eq 20.3 in [1], or Eq 13 in [2].
-
-    # Arguments:
-        Wx: np.ndarray
-            CWT of `x` (see `cwt`).
-
-        dWx: np.ndarray.
-            Time-derivative of `Wx`, computed via frequency-domain differentiation
-            (effectively, derivative of trigonometric interpolation; see [4]).
-
-        difftype: str['direct', 'phase']
-            Method by which to differentiate Wx (default='direct') to obtain
-            instantaneous frequencies:
-                    w(a,b) = Im( (1/2pi) * (1/Wx(a,b)) * d/db[Wx(a,b)] )
-
-                - 'direct': using `dWx` (see `dWx`).
-                - 'phase': differentiate by taking forward finite-difference of
-                unwrapped angle of `Wx` (see `phase_cwt`).
-
-        gamma: float / None
-            CWT phase threshold. Sets `w=inf` for small values of `Wx` where
-            phase computation is unstable and inaccurate (like in DFT):
-                w[abs(Wx) < beta] = inf
-            This is used to zero `Wx` where `w=0` in computing `Tx` to ignore
-            contributions from points with indeterminate phase.
-            Default = sqrt(machine epsilon) = np.sqrt(np.finfo(np.float64).eps)
-
-    # Returns:
-        w: np.ndarray
-            Phase transform for each element of `Wx`. w.shape == Wx.shape.
-
-    # References:
-        1. A Nonlinear squeezing of the CWT Based on Auditory Nerve Models.
-        I. Daubechies, S. Maes.
-        https://services.math.duke.edu/%7Eingrid/publications/DM96.pdf
-
-        2. The Synchrosqueezing algorithm for time-varying spectral analysis:
-        robustness properties and new paleoclimate applications.
-        G. Thakur, E. Brevdo, N.-S. Fučkar, and H.-T. Wu.
-        https://arxiv.org/abs/1105.0010
-
-        3. Synchrosqueezed Wavelet Transforms: a Tool for Empirical Mode
-        Decomposition. I. Daubechies, J. Lu, H.T. Wu.
-        https://arxiv.org/pdf/0912.2437.pdf
-
-        4. The Exponential Accuracy of Fourier and Chebyshev Differencing Methods.
-        E. Tadmor.
-        http://webhome.auburn.edu/~jzl0097/teaching/math_8970/Tadmor_86.pdf
-
-        5. Synchrosqueezing Toolbox, (C) 2014--present. E. Brevdo, G. Thakur.
-        https://github.com/ebrevdo/synchrosqueezing/blob/master/synchrosqueezing/
-        phase_cwt.m
-    """
-    if difftype == 'phase':
-        # TODO gives bad results; shouldn't we divide by Wx?
-        u = np.unwrap(np.angle(Wx)).T
-        w = np.vstack([np.diff(u, axis=0), u[-1] - u[0]]).T / (2*pi)
-    else:
-        with np.errstate(divide='ignore'):
-            w = np.imag(dWx / Wx) / (2*pi)
-    # treat negative phases as positive; these are in small minority, and
-    # slightly aid invertibility (as less of `Wx` is zeroed in ssqueezing)
-    w = np.abs(w)
-
-    gamma = gamma or np.sqrt(EPS)
-    w[np.abs(Wx) < gamma] = np.inf
-    return w
-
-
-def phase_cwt_num(Wx, dt, difforder=4, gamma=None):
-    """Calculate the phase transform at each (scale, time) pair:
-        w[a, b] = Im((1/2pi) * d/db (Wx[a,b]) / Wx[a,b])
-    Uses numeric differentiation (1st, 2nd, or 4th order). See above Eq 20.3
-    in [1], or Eq 13 in [2].
-
-    # Arguments:
-        Wx: np.ndarray. Wavelet transform of `x` (see `cwt`).
-
-        dt: int. Sampling period (e.g. t[1] - t[0]).
-
-        difforder: int[1, 2, 4]
-            Order of differentiation (default=4).
-
-        gamma: float
-            CWT phase threshold. Sets `w=inf` for small values of `Wx` where
-            phase computation is unstable and inaccurate (like in DFT):
-                w[abs(Wx) < beta] = inf
-            This is used to zero `Wx` where `w=0` in computing `Tx` to ignore
-            contributions from points with indeterminate phase.
-            Default = sqrt(machine epsilon) = np.sqrt(np.finfo(np.float64).eps)
-
-    # Returns:
-        w: np.ndarray
-            Phase transform via demodulated FM-estimates. w.shape == Wx.shape.
-
-    # References:
-        1. A Nonlinear squeezing of the CWT Based on Auditory Nerve Models.
-        I. Daubechies, S. Maes.
-        https://services.math.duke.edu/%7Eingrid/publications/DM96.pdf
-
-        2. The Synchrosqueezing algorithm for time-varying spectral analysis:
-        robustness properties and new paleoclimate applications.
-        G. Thakur, E. Brevdo, N.-S. Fučkar, and H.-T. Wu.
-        https://arxiv.org/abs/1105.0010
-
-        3. Synchrosqueezing Toolbox, (C) 2014--present. E. Brevdo, G. Thakur.
-        https://github.com/ebrevdo/synchrosqueezing/blob/master/synchrosqueezing/
-        phase_cwt_num.m
-    """
-    # unreliable; bad results on high freq pure tones
-    def _differentiate(Wx, dt):
-        if difforder in (2, 4):
-            # append for differentiating
-            Wxr = np.hstack([Wx[:, -2:], Wx, Wx[:, :2]])
-
-        if difforder == 1:
-            w = np.hstack([Wx[:, 1:] - Wx[:, :-1],
-                           Wx[:, :1]  - Wx[:, -1:]])
-            w /= dt
-        elif difforder == 2:
-            # calculate 2nd-order forward difference
-            w = -Wxr[:, 4:] + 4 * Wxr[:, 3:-1] - 3 * Wxr[:, 2:-2]
-            w /= (2 * dt)
-        elif difforder == 4:
-            # calculate 4th-order central difference
-            w = -Wxr[:, 4:]
-            w += Wxr[:, 3:-1] * 8
-            w -= Wxr[:, 1:-3] * 8
-            w += Wxr[:, 0:-4]
-            w /= (12 * dt)
-        return w
-
-    # epsilon from Daubechies, H-T Wu, et al.
-    # gamma from Brevdo, H-T Wu, et al.
-    gamma = gamma or np.sqrt(EPS)
-    if difforder not in (1, 2, 4):
-        raise ValueError("`difforder` must be one of: 1, 2, 4 "
-                         "(got %s)" % difforder)
-
-    w = _differentiate(Wx, dt)
-    w[np.abs(Wx) < gamma] = np.inf
-
-    # calculate inst. freq for each scale
-    # 2*pi norm per discretized inverse FT rather than inverse DFT
-    w = np.real(-1j * w / Wx) / (2*pi)
-    w[w < 0] = np.inf
-    return w
+def _check_squeezing_arg(squeezing):
+    supported = ('sum', 'lebesgue', 'abs')
+    if not isinstance(squeezing, (str, FunctionType)):
+        raise TypeError("`squeezing` must be string or function "
+                        "(got %s)" % type(squeezing))
+    elif isinstance(squeezing, str) and squeezing not in supported:
+        raise ValueError("`squeezing` must be one of: {} (got {})".format(
+            ', '.join(supported), squeezing))
